+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 21.10.2
 - more issues with the translation of the default buttons under macOS were
   tried to be fixed (for [#783](https://github.com/pbek/QOwnNotes/issues/783))
+- the restart information in single app instance mode was improved to make it
+  more understandable (for [#2235](https://github.com/pbek/QOwnNotes/issues/2235))
 
 ## 21.10.1
 - when copying or moving notes to another sub-folder now notes will not be copied